reuse dynamic check
