+++ src/tests/functional/plugin/conformance/subgraphs_dumper/include/utils/model.hpp
@@ -18,6 +18,7 @@
 
 #include "cache/cache.hpp"
 #include "utils/node.hpp"
+#include "utils/dynamism.hpp"
 
 namespace ov {
 namespace tools {
@@ -75,7 +76,6 @@ std::map<ModelCacheStatus, std::vector<std::string>> cache_models(
 void save_model_status_to_file(const std::map<ModelCacheStatus, std::vector<std::string>>& caching_status,
                                const std::string& output_dir);
 
-bool is_dynamic_model(const std::shared_ptr<ov::Model>& model);
 std::string get_model_type(const std::shared_ptr<ov::Model>& model);
 
 std::map<std::string, InputInfo>
